rename: RssChecker 이름을 RssSubscribeParser로 변경 (#193)
- 문서를 체크하는 기능 보다 문서를 파싱해서 파싱한 데이터를 가공해서 반환하는 기능이 주요 기능이므로 변경
